v0.12.5 - Add --ignore-missing-imports to mypycify
Not sure why, but it seems like mypycify checks the files before
the requirements are installed, leading to errors otherwise.
So we need to pass this flag into mypycify to supress this behavior.
Previously I used # type: ignore comments on all of the imports but
removing these yields better performance since mypy
can type check the imports, plus this is more elegant as well.